+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Added `sparv plugins` command to list and manage Sparv plugins.
 - Added `--compact` flag for `sparv schema` command, resulting in compact output.
 - Added `--json` flag for `sparv modules` command.
 - The command `sparv modules` now also takes annotators in addition to modules as arguments to filter the output.